moved types files to a separate folder, installed zod, converted types to zod schemas
