--- src/components/foh-metabox-temporary-types.ts
@@ -1,6 +1,0 @@
-import { DatePickerEvent } from "@wordpress/components/build-types/date-time/types"
-
-export type DatesRange = {
-    start: DatePickerEvent
-    end: DatePickerEvent
-}
